rename component tsx files.
