+++ tic-tac-toe-app/src/index.tsx
@@ -1,7 +1,7 @@
 import ReactDOM from "react-dom";
 import "./index.css";
 import reportWebVitals from "./reportWebVitals";
-import { Game } from "./components/game";
+import { Game } from "./components/game/Game";
 // ========================================
 
 ReactDOM.render(<Game />, document.getElementById("root"));
